deduplicate safe_call
