+++ droidlet/lowlevel/hello_robot/hello_robot_mover.py
@@ -33,6 +33,7 @@ from droidlet.lowlevel.hello_robot.rotation import (
 )
 
 from tenacity import retry, stop_after_attempt, wait_fixed
+from droidlet.lowlevel.pyro_utils import safe_call
 
 Pyro4.config.SERIALIZER = "pickle"
 Pyro4.config.SERIALIZERS_ACCEPTED.add("pickle")
@@ -41,17 +42,6 @@ Pyro4.config.PICKLE_PROTOCOL_VERSION=2
 MAX_PAN_RAD = math.pi/4
 
 # TODO/FIXME: state machines.  state machines everywhere
-
-def safe_call(f, *args, **kwargs):
-    try:
-        return f(*args, **kwargs)
-    except Pyro4.errors.ConnectionClosedError as e:
-        msg = "{} - {}".format(f._RemoteMethod__name, e)
-        raise ErrorWithResponse(msg)
-    except Exception as e:
-        print("Pyro traceback:")
-        print("".join(Pyro4.util.getPyroTraceback()))
-        raise e
 
 class HelloRobotMover(MoverInterface):
     """Implements methods that call the physical interfaces of the Robot.
